+++ ipc/socket/vsock/samples/config.h
@@ -1,8 +1,4 @@
 #pragma once
-#include <sys/socket.h>
-#include <linux/vm_sockets.h>
 
 #define PORT	6789
-#define CID_HOST	VMADDR_CID_ANY
-#define CID_GUEST	VMADDR_CID_HOST