+++ cpp/fmt.cpp
@@ -1,0 +1,10 @@
+/**
+ * Test https://github.com/fmtlib/fmt
+ */
+#include <fmt/base.h>
+
+int main(void)
+{
+	fmt::print("Hello, world!\n");
+	return 0;
+}