+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to
 ## Unreleased
 #### Breaking Changes
 #### Added
+- kata-containers: First commit.
 - procfs: add mounts.sh.
 - cpp: try test fmtlib.
 - cgroup: Support sub-dir.